deteccion_objetos/requirements.txt
Django==5.2.7
opencv-python==4.12.0.88
Pillow==12.0.0
numpy==2.2.6
asgiref==3.10.0
sqlparse==0.5.3
tzdata==2025.2
